Update Chrono24_scraper.ipynb
Forgot to add most important attribute of all: Price.

Price is now being fetched and added to the attributes dictionary.

There are 2 different tags for price.

Main:

span ->"class":'js-price-shipping-country'

If this tag is non existent or empty, then fetch content from:

span->"class":'price-md'

If the price on the ad is "Price on request", this will be the text that's populated in the main tag and therefore no adjustment is needed