Brand selection refinement
While the main goal was to extract all the ads for all the brands, this proved to be very slow (specially on a VPN), so the list of brands was trimmed to only include the top ones.

PS: Rolex accounts for a bit under 1/3 (~70k) of all the ads provided by the new brand list. It's recommended to fetch the details for Rolex Ads separately from the rest